Add UiKitScreen component for Ui Kit demo
