+++ modules/app_ui_kit_demo/src/main/kotlin/kekmech/ru/mpeiapp/demo/screens/typography/TypographyScreen.kt
@@ -1,18 +1,11 @@
 package kekmech.ru.mpeiapp.demo.screens.typography
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.nestedscroll.nestedScroll
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
 import com.bumble.appyx.core.node.node
 import kekmech.ru.common_navigation_api.NavTarget
-import kekmech.ru.common_navigation_compose.LocalBackStackNavigator
-import kekmech.ru.ui_kit_topappbar.TopAppBar
+import kekmech.ru.mpeiapp.demo.ui.UiKitScreen
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -25,24 +18,7 @@ internal class TypographyScreenNavTarget : NavTarget {
 
 @Composable
 private fun TypographyScreen() {
-    val navigator = LocalBackStackNavigator.current
-    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
-
-    Scaffold(
-        modifier = Modifier
-            .nestedScroll(scrollBehavior.nestedScrollConnection),
-        topBar = {
-            TopAppBar(
-                title = "Typography",
-                scrollBehavior = scrollBehavior,
-                navigationIcon = {
-                    BackIconButton {
-                        navigator.back()
-                    }
-                }
-            )
-        },
-    ) { innerPadding ->
-        Box(modifier = Modifier.padding(innerPadding))
+    UiKitScreen(title = "Typography") {
+        /* no-op */
     }
 }
